Move category search to search module
