+++ indico/modules/categories/views.py
@@ -96,16 +96,3 @@ class WPCategoryManagement(WPCategory):
 
 class WPCategoryStatistics(WPCategory):
     bundles = ('module_categories.css',)
-
-
-class WPCategorySearch(WPCategory):
-    """WP for category-limited search"""
-    bundles = ('module_search.js', 'module_search.css')
-
-    def __init__(self, rh, category, **kwargs):
-        WPCategory.__init__(self, rh, category, **kwargs)
-
-    def _get_breadcrumbs(self):
-        if not self.category or self.category.is_root:
-            return ''
-        return render_breadcrumbs(_('Search'), category=self.category)
