+++ indico/modules/search/templates/category_search.html
